+++ python/metatensor-learn/metatensor/learn/nn/silu.py
@@ -4,6 +4,7 @@ import torch
 from torch.nn import Module
 
 from .._backend import Labels, TensorMap
+from .._dispatch import int_array_like
 from .module_map import ModuleMap
 
 
@@ -33,7 +34,7 @@ class SiLU(Module):
         in_place: bool = False,
     ) -> None:
         super().__init__()
-        modules: List[Module] = [torch.nn.SiLU() for i in range(len(in_keys))]
+        modules: List[Module] = [torch.nn.SiLU(in_place) for i in range(len(in_keys))]
         self.module_map = ModuleMap(in_keys, modules, out_properties)
 
     def forward(self, tensor: TensorMap) -> TensorMap:
@@ -62,35 +63,43 @@ class InvariantSiLU(torch.nn.Module):
 
     Applies a sigmoid linear unit transformation to each invariant block of a
     :py:class:`TensorMap` passed to its :py:meth:`forward` method. These are indexed by
-    the keys in :param in_keys: at numeric indices passed in :param invariant_key_idxs:.
+    the keys in :param in_keys: that correspond to the selection passed in :param
+    invariant_keys:.
 
     Refer to the :py:class`torch.nn.SiLU` documentation for a more detailed description
     of the parameters.
 
     :param in_keys: :py:class:`Labels`, the keys that are assumed to be in the input
         tensor map in the :py:meth:`forward` method.
-    :param invariant_key_idxs: list of int, the indices of the invariant keys present in
-        `in_keys` in the input :py:class:`TensorMap`. Only blocks for these keys will
-        have the SiLU transformation applied. Covariant blocks will have the identity
-        operator applied.
     :param out_properties: list of :py:class`Labels` (optional), the properties labels
         of the output. By default the output properties are relabeled using
         Labels.range.
+    :param invariant_keys: a :py:class:`Labels` object that is used to select the
+        invariant keys from ``in_keys``. If not provided, the invariant keys are assumed
+        to be those where key dimensions ``["o3_lambda", "o3_sigma"]`` are equal to
+        ``[0, 1]``.
     """
 
     def __init__(
         self,
         in_keys: Labels,
-        invariant_key_idxs: List[int],
         out_properties: Optional[Labels] = None,
+        invariant_keys: Optional[Labels] = None,
         *,
         in_place: bool = False,
     ) -> None:
         super().__init__()
+        # Set a default for invariant keys
+        if invariant_keys is None:
+            invariant_keys = Labels(
+                names=["o3_lambda", "o3_sigma"],
+                values=int_array_like([0, 1], like=in_keys.values).reshape(-1, 1),
+            )
+        invariant_key_idxs = in_keys.select(invariant_keys)
         modules: List[Module] = []
         for i in range(len(in_keys)):
             if i in invariant_key_idxs:  # Invariant block: apply SiLU
-                module = torch.nn.SiLU()
+                module = torch.nn.SiLU(in_place)
             else:  # Covariant block: apply identity operator
                 module = torch.nn.Identity()
             modules.append(module)